moved structs to thea primitives
