+++ thea-primitives/Cargo.toml
@@ -11,6 +11,7 @@ blst = {version = "0.3.10", default-features = false, optional=true}
 frame-system = { workspace = true, default-features = false }
 polkadex-primitives = {workspace = true, default-features = false}
 sp-std = { workspace = true, default-features = false }
+sp-core = { workspace = true, default-features = false }
 
 # For primitive structs
 parity-scale-codec = { workspace = true, default-features = false, features = ["max-encoded-len"] }
@@ -24,4 +25,5 @@ std = [
     "frame-system/std",
     "parity-scale-codec/std",
     "scale-info/std",
+    "sp-core/std",
 ]
